% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_join_data.R
\name{add_outcome_variable}
\alias{add_outcome_variable}
\title{Add outcome variable}
\usage{
add_outcome_variable(read_counts, sample_names)
}
\arguments{
\item{read_counts}{A list of read counts of all the samples}
}
\value{
A dataframe with an outcome variable added
}
\description{
Add outcome variable
}
