% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_estimate_sample_index_hopping_rate.R
\name{create_chimera_counts}
\alias{create_chimera_counts}
\title{Create chimera counts datatable}
\usage{
create_chimera_counts(outcome_counts, S)
}
\arguments{
\item{outcome_counts}{outcome dataset}

\item{S}{number of samples}
}
\value{
Dataframe of chimera counts
}
\description{
Create chimera counts datatable
}
