% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_create_counts_by_outcome_table.R
\name{create_grouping_vars}
\alias{create_grouping_vars}
\title{Create grouping variables}
\usage{
create_grouping_vars(outcome_counts, sample_names)
}
\arguments{
\item{outcome_counts}{outcome dataset}

\item{sample_names}{sample names}
}
\value{
dataframe of grouping variables
}
\description{
Create grouping variables
}
