% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_estimate_sample_index_hopping_rate.R
\name{fit_glm}
\alias{fit_glm}
\title{Estimate hopping rate}
\usage{
fit_glm(chimera_counts, S, max_r, conf_level = 0.99)
}
\arguments{
\item{max_r}{Maximum PCR duplication level to consider}

\item{outcome_counts}{outcome dataset}
}
\value{
list of two dataframes (glm estimates and chimera counts)
}
\description{
Estimate hopping rate
}
