% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_create_joined_counts_table.R
\name{get_joined_counts_table}
\alias{get_joined_counts_table}
\title{Join read counts of all samples}
\usage{
get_joined_counts_table(read_counts = NULL, joined_counts_filepath = NULL)
}
\arguments{
\item{read_counts}{A list of read counts of all the samples}

\item{joined_counts_filepath}{If provided, the filepath of the joined readcounts}
}
\value{
A joined dataframe of read counts and an outcome variable
}
\description{
Join read counts of all samples
}
