% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_create_joined_counts_table.R
\name{join_read_counts}
\alias{join_read_counts}
\title{Join transciptome mapped data by cell, umi, and gene}
\usage{
join_read_counts(read_counts, keys = c("cell", "gene", "umi"))
}
\arguments{
\item{read_counts}{A list of read counts of all the samples}

\item{keys}{keys for joining}
}
\value{
A joined read counts table
}
\description{
Join transciptome mapped data by cell, umi, and gene
}
