% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_purge_data.R
\name{load_and_estimate_sihr}
\alias{load_and_estimate_sihr}
\title{Compute summary statistics}
\usage{
load_and_estimate_sihr(samples, barcode_length = NULL, max_r = NULL)
}
\arguments{
\item{samples}{samples}

\item{max_r}{Maximum PCR duplication level to consider}

\item{barcode.length}{the length of the cell barcode (for v2 data)}
}
\value{
list
}
\description{
Compute summary statistics
}
