% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_create_joined_counts_table.R
\name{load_molecule_info_data}
\alias{load_molecule_info_data}
\title{Load CellRanger data  for all samples}
\usage{
load_molecule_info_data(input_dir)
}
\arguments{
\item{input_dir}{The filepath of the directory}
}
\value{
A dataframe of read counts for all samples
}
\description{
Load CellRanger data  for all samples
}
