% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_purge_data.R
\name{make_count_matrices}
\alias{make_count_matrices}
\title{make sparse count matrices}
\usage{
make_count_matrices(umi_counts, all_genes, return_discarded = TRUE)
}
\arguments{
\item{umi_counts}{list of sample data}

\item{all_genes}{ref_genes}

\item{return_discarded}{return discarded data}
}
\value{
list of two lists containing retained and discarded data
}
\description{
make sparse count matrices
}
