% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhantomPurgeR.R
\docType{package}
\name{PhantomPurgeR}
\alias{PhantomPurgeR}
\title{Workflow: Phantom Purge
\enumerate{
\item  \code{\link{load_molecule_info_data}}
loads molecule_info files and creates read counts datatable.
Each sample's \emph{molecule_info.h5} file should be renamed to
\emph{{sample_name}.h5} and placed in the input_dir folder.
\item  \code{\link{get_joined_counts_table}}
join read counts of all samples
\item  \code{\link{create_outcome_counts}}
create outcome counts datatable
\item  \code{\link{estimate_hopping_rate}}
create a chimera counts datatable and estimate hopping rate
}}
\description{
Workflow: Phantom Purge
\enumerate{
\item  \code{\link{load_molecule_info_data}}
loads molecule_info files and creates read counts datatable.
Each sample's \emph{molecule_info.h5} file should be renamed to
\emph{{sample_name}.h5} and placed in the input_dir folder.
\item  \code{\link{get_joined_counts_table}}
join read counts of all samples
\item  \code{\link{create_outcome_counts}}
create outcome counts datatable
\item  \code{\link{estimate_hopping_rate}}
create a chimera counts datatable and estimate hopping rate
}
}
