% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_plotting_functions.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot model fit}
\usage{
plot_fit(data_list, dataset_name, x_lim = 150)
}
\arguments{
\item{data_list}{output list}

\item{dataset_name}{name of dataset for plotting}

\item{x_lim}{extent of x axis}
}
\value{
ggplot object
}
\description{
Plot model fit
}
