% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_molecules_distributions}
\alias{plot_molecules_distributions}
\title{Plot molecular proportions complexityprofile}
\usage{
plot_molecules_distributions(data_list, dataset_name, x_lim = 150)
}
\arguments{
\item{data_list}{output list}

\item{dataset_name}{name of dataset for plotting}

\item{x_lim}{extent of x axis}
}
\value{
ggplot object
}
\description{
Plot molecular proportions complexityprofile
}
