% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_reassign_hopped_reads.R
\name{purge_and_save_read_counts}
\alias{purge_and_save_read_counts}
\title{Purge and save read_counts}
\usage{
purge_and_save_read_counts(read_counts, dataset_name, sample_names, output_dir)
}
\arguments{
\item{read_counts}{read counts}

\item{dataset_name}{dataset name}

\item{sample_names}{sample names}

\item{output_dir}{output dir}
}
\value{

}
\description{
Purge and save read_counts
}
