% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_purge_data.R
\name{purge_data}
\alias{purge_data}
\title{Reassign reads}
\usage{
purge_data(output, torc, get_discarded = TRUE)
}
\arguments{
\item{output}{Output of estimate_hopping_rate function}

\item{torc}{TOR cutoff}

\item{read_counts}{read counts}

\item{outcome_counts}{outcome dataset}
}
\value{
outcome dataset
}
\description{
Reassign reads
}
