% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_create_joined_counts_table.R
\name{read10xMolInfo_2}
\alias{read10xMolInfo_2}
\title{Load CellRanger 10xv2 data
The function uses cxx_get_cell_barcodes from dropUtils}
\usage{
read10xMolInfo_2(sample)
}
\arguments{
\item{sample}{The name of the filename (i.e. sample)}
}
\value{
A dataframe
}
\description{
Load CellRanger 10xv2 data
The function uses cxx_get_cell_barcodes from dropUtils
}
