% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_create_joined_counts_table.R
\name{read10xMolInfo_3}
\alias{read10xMolInfo_3}
\title{Load CellRanger 10xv3 data}
\usage{
read10xMolInfo_3(sample)
}
\arguments{
\item{sample}{The name of the filename (i.e. sample)}
}
\value{
A dataframe
}
\description{
Load CellRanger 10xv3 data
}
