% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_create_joined_counts_table.R
\name{read10xMolInfoSamples}
\alias{read10xMolInfoSamples}
\title{Load data from molecule_info.h5 files produced by 10X Genomics CellRanger software.}
\usage{
read10xMolInfoSamples(samples, barcode_length = NULL)
}
\arguments{
\item{samples}{A named list of filepaths}

\item{barcode_length}{the length of the cell barcode for v2 data}
}
\value{
lists of cell, gene, umi, read count, and ref_genes variables
}
\description{
Load data from molecule_info.h5 files produced by 10X Genomics CellRanger software.
}
