% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_reassign_hopped_reads.R
\name{reassign_reads_and_mark_retained_observations}
\alias{reassign_reads_and_mark_retained_observations}
\title{Reassign reads}
\usage{
reassign_reads_and_mark_retained_observations(
  read_counts,
  outcome_counts,
  fit_out,
  torc
)
}
\arguments{
\item{read_counts}{read counts}

\item{outcome_counts}{outcome dataset}

\item{fit_out}{Output of estimate_hopping_rate function}

\item{torc}{TOR cutoff}
}
\value{
outcome dataset
}
\description{
Reassign reads
}
