% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_create_joined_counts_table.R
\name{rename_var_data_list}
\alias{rename_var_data_list}
\title{Rename variables}
\usage{
rename_var_data_list(read_counts, sample_names)
}
\arguments{
\item{read_counts}{A list of read counts of all the samples}
}
\value{
A joined read counts table
}
\description{
Rename variables
}
