% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_estimate_hopping_rate.R
\name{update_chimera_counts}
\alias{update_chimera_counts}
\title{Update chimera counts data}
\usage{
update_chimera_counts(chimera_counts, glm_estimates)
}
\arguments{
\item{chimera_counts}{A dataframe of chimera counts}

\item{glm_estimates}{A dataframe of GLM estimates}
}
\value{
A dataframe of chimera counts
}
\description{
Update chimera counts data
}
