% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_estimate_hopping_rate.R
\name{add_vars_to_outcome_counts}
\alias{add_vars_to_outcome_counts}
\title{Add grouping variables}
\usage{
add_vars_to_outcome_counts(outcome_counts, sample_names)
}
\arguments{
\item{outcome_counts}{outcome dataset}

\item{sample_names}{sample names}
}
\value{
Datatable of outcome counts
}
\description{
Add grouping variables
}
