% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_purge_data.R
\name{compute_summary_stats}
\alias{compute_summary_stats}
\title{Compute summary statistics}
\usage{
compute_summary_stats(outcome_counts, phat)
}
\arguments{
\item{outcome_counts}{outcome dataset}

\item{max_r}{Maximum PCR duplication level to consider}
}
\value{
list containing two dataframes (glm estimates and chimera counts)
}
\description{
Compute summary statistics
}
