% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_estimate_hopping_rate.R
\name{create_outcome_counts}
\alias{create_outcome_counts}
\title{Create outcome counts}
\usage{
create_outcome_counts(read_counts, sample_names)
}
\arguments{
\item{read_counts}{A list of read counts of all the samples}

\item{joined_counts_filepath}{If provided, the filepath of the joined readcounts to be saved}
}
\value{
Datatable of outcome counts
}
\description{
Create outcome counts
}
