% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_purge_phantom_molecules.R
\name{create_umi_counts}
\alias{create_umi_counts}
\title{Purge and save read_counts}
\usage{
create_umi_counts(read_counts, sample_names)
}
\arguments{
\item{read_counts}{read counts}

\item{sample_names}{sample names}
}
\value{
umi_counts_cell_gene dataset
}
\description{
Purge and save read_counts
}
