% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_estimate_sample_index_hopping_rate.R
\name{estimate_hopping_rate}
\alias{estimate_hopping_rate}
\title{Estimate hopping rate}
\usage{
estimate_hopping_rate(outcome_counts, max_r = NULL)
}
\arguments{
\item{outcome_counts}{outcome dataset}

\item{max_r}{Maximum PCR duplication level to consider}
}
\value{
list containing two dataframes (glm estimates and chimera counts)
}
\description{
Estimate hopping rate
}
