% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_estimate_sample_index_hopping_rate.R
\name{estimate_hopping_rate}
\alias{estimate_hopping_rate}
\title{Estimate hopping rate}
\usage{
estimate_hopping_rate(out, max_r = NULL)
}
\arguments{
\item{out}{list}

\item{max_r}{Maximum PCR duplication level to consider}
}
\value{
out list containing three additional dataframes (outcome_counts, glm estimates and chimera counts)
}
\description{
Estimate hopping rate
}
