% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_estimate_hopping_rate.R
\name{estimate_hopping_rate}
\alias{estimate_hopping_rate}
\title{Estimate hopping rate}
\usage{
estimate_hopping_rate(out, max_r = NULL)
}
\arguments{
\item{out}{list containing read_counts and sample_names}

\item{max_r}{Maximum PCR duplication level to consider}
}
\value{
list containing four dataframes: outcome_counts, glm estimates, chimera counts, and summary_stats
}
\description{
Estimate hopping rate
}
