% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_load_data.R
\name{get_h5_filenames}
\alias{get_h5_filenames}
\title{Return filepaths of CellRanger molecule_info.h5 files in directory}
\usage{
get_h5_filenames(input_dir)
}
\arguments{
\item{input_dir}{The directory in which the molecule_info.h5 files are located}
}
\value{
samples_filepaths: A named list of samples filepaths
}
\description{
Return filepaths of CellRanger molecule_info.h5 files in directory
}
