% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_joined_data.R
\name{join_data}
\alias{join_data}
\title{Join transciptome mapped data by cell, umi, and gene}
\usage{
join_data(out)
}
\arguments{
\item{A}{list of read counts data for all samples}
}
\value{
A  joined read counts table with an outcome variable column
}
\description{
Join transciptome mapped data by cell, umi, and gene
}
