% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_create_joined_counts_table.R
\name{join_read_counts}
\alias{join_read_counts}
\title{Join transciptome mapped data by cell, umi, and gene}
\usage{
join_read_counts(out)
}
\arguments{
\item{A}{list of read counts data for all samples}
}
\value{
A  joined read counts table with an outcome variable column
}
\description{
Join transciptome mapped data by cell, umi, and gene
}
