% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plot_data.R
\name{make_plots}
\alias{make_plots}
\title{Make plots}
\usage{
make_plots(out, dataset_name, x_lim = 160, legend_rel_width = 0.2)
}
\arguments{
\item{out}{output list}

\item{dataset_name}{name of dataset for plotting}

\item{x_lim}{extent of x axis}

\item{legend_rel_width}{relative width of legend subplot}
}
\value{
list of ggplot objects
}
\description{
Make plots
}
