% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_purge_data.R
\name{phantom_purger}
\alias{phantom_purger}
\title{Run Purge Phantoms workflow}
\usage{
phantom_purger(
  samples,
  torc,
  max_r = NULL,
  barcode_length = NULL,
  return_readcounts = FALSE,
  return_discarded = TRUE
)
}
\arguments{
\item{samples}{A named list of filepaths}

\item{torc}{TOR cutoff}

\item{max_r}{Maximum PCR duplication level to consider}

\item{barcode_length}{the length of the cell barcode for v2 data}

\item{return_readcounts}{If true the joined readcounts is returned}

\item{return_discarded}{return discarded data}
}
\value{
out list pf umi_counts, metadata, and summary statistics
}
\description{
Run Purge Phantoms workflow
}
