% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhantomPurgeR.R
\docType{package}
\name{PhantomPurgeR}
\alias{PhantomPurgeR}
\title{Workflow: Phantom Purge
\enumerate{
\item  \code{\link{get_h5_filenames}}
Returns named list of sample filepaths of CellRanger molecule_info.h5 files in provided directory.
\item  \code{\link{read10xMolInfoSamples}}
Loads molecule_info.h5 samples. Renamed files in the input_dir folder as
\emph{{sample_name}.h5} or rename the list of samples' filepaths.
\item  \code{\link{join_data}}
Join read counts of all samples. Merge mapped data by cell, umi, and gene keys.
\item  \code{\link{estimate_hopping_rate}}
Create outcome counts datatable, chimera counts datatable, and estimate the sample index hopping rate.
\item  \code{\link{purge_data}}
Reassign hopped reads and purge phantom molecules.
\item  \code{\link{make_plots}}
Create diagnostics plots.
}}
\description{
Workflow: Phantom Purge
\enumerate{
\item  \code{\link{get_h5_filenames}}
Returns named list of sample filepaths of CellRanger molecule_info.h5 files in provided directory.
\item  \code{\link{read10xMolInfoSamples}}
Loads molecule_info.h5 samples. Renamed files in the input_dir folder as
\emph{{sample_name}.h5} or rename the list of samples' filepaths.
\item  \code{\link{join_data}}
Join read counts of all samples. Merge mapped data by cell, umi, and gene keys.
\item  \code{\link{estimate_hopping_rate}}
Create outcome counts datatable, chimera counts datatable, and estimate the sample index hopping rate.
\item  \code{\link{purge_data}}
Reassign hopped reads and purge phantom molecules.
\item  \code{\link{make_plots}}
Create diagnostics plots.
}
}
