% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhantomPurgeR.R
\docType{package}
\name{PhantomPurgeR}
\alias{PhantomPurgeR}
\title{PhantomPurgeR Workflow}
\details{
See workflow vignettes on package's \href{https://csglab.github.io/PhantomPurgeR/pages/rpackage.html}{website}.

Description of the workflow steps.
\enumerate{
\item  \code{\link{get_h5_filenames}}
\item  \code{\link{read10xMolInfoSamples}}
\item  \code{\link{join_data}}
\item  \code{\link{estimate_hopping_rate}}
\item  \code{\link{purge_phantoms}}
\item  \code{\link{make_plots}}
}
}
