% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_tor}
\alias{plot_tor}
\title{Plot TOR curves}
\usage{
plot_tor(out, dataset_name)
}
\arguments{
\item{out}{output list}

\item{dataset_name}{name of dataset for plotting}

\item{x_lim}{extent of x axis}
}
\value{
ggplot object
}
\description{
Plot TOR curves
}
