% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_purge_data.R
\name{purge_phantoms}
\alias{purge_phantoms}
\title{Compute summary statistics}
\usage{
purge_phantoms(
  samples,
  torc = 3,
  return_readcounts = FALSE,
  barcode_length = NULL,
  max_r = NULL
)
}
\arguments{
\item{samples}{samples}

\item{torc}{torc}

\item{return_readcounts}{If true the joined readcounts is returned}

\item{max_r}{Maximum PCR duplication level to consider}

\item{barcode.length}{the length of the cell barcode (for v2 data)}
}
\value{
list
}
\description{
Compute summary statistics
}
