% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_purge_data.R
\name{purge_phantoms}
\alias{purge_phantoms}
\title{Reassign reads}
\usage{
purge_phantoms(out, torc, return_readcounts = FALSE, return_discarded = TRUE)
}
\arguments{
\item{out}{out from previous steps}

\item{torc}{TOR cutoff}

\item{return_readcounts}{If true the joined readcounts is returned}

\item{return_discarded}{return discarded data}
}
\value{
out list with additional objects; umi_counts and summary_stats
}
\description{
Reassign reads
}
